requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
python-dateutil==2.8.2
ratelimit==2.2.1
